// src/types/enums/IntervalType.ts
export enum IntervalType {
  Minor2 = "Interval_Min2",
  Major2 = "Interval_Maj2",
  Minor3 = "Interval_Min3",
  Major3 = "Interval_Maj3",
  Fourth = "Interval_Fourth",
  Tritone = "Interval_Tritone",
  Fifth = "Interval_Fifth",
  Minor6 = "Interval_Min6",
  Major6 = "Interval_Maj6",
  Minor7 = "Interval_Min7",
  Major7 = "Interval_Maj7",
  Octave = "Interval_Octave",
}